Move all helpers to helpers namespace
