+++ ruby/lib/cucumber/messages/helpers/id_generator.rb
