+++ ruby/lib/cucumber/messages/helpers/id_generator/uuid.rb
