+++ ruby/lib/cucumber/messages/helpers/ndjson_to_message_enumerator.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+require 'cucumber/messages.deserializers'
+
+module Cucumber
+  module Messages
+    module Helpers
+      class NdjsonToMessageEnumerator < Enumerator
+        def initialize(io)
+          super() do |yielder|
+            io.each_line do |line|
+              next if line.strip.empty?
+              begin
+                m = Envelope.from_json(line)
+              rescue StandardError
+                raise "Not JSON: #{line.strip}"
+              end
+              yielder.yield(m)
+            end
+          end
+        end
+      end
+    end
+  end
+end
