--- ruby/lib/cucumber/messages/ndjson_to_message_enumerator.rb
@@ -1,23 +1,0 @@
-# frozen_string_literal: true
-
-require 'cucumber/messages.deserializers'
-
-module Cucumber
-  module Messages
-    class NdjsonToMessageEnumerator < Enumerator
-      def initialize(io)
-        super() do |yielder|
-          io.each_line do |line|
-            next if line.strip.empty?
-            begin
-              m = Envelope.from_json(line)
-            rescue StandardError
-              raise "Not JSON: #{line.strip}"
-            end
-            yielder.yield(m)
-          end
-        end
-      end
-    end
-  end
-end
